Run Forge Format

+++ apps/forge/script/CommitAndRegister.s.sol
@@ -29,18 +29,12 @@ contract Template is Script, DeployFNSAbstract {
         vm.startBroadcast(broadcastPrivKey);
 
         bytes32 commitment = flrRegistrarController.makeCommitment(
-            name,
-            broadcastAddress,
-            31556952,
-            keccak256(bytes(name)),
-            address(publicResolver),
-            new bytes[](0),
-            true,
-            0);
-        
+            name, broadcastAddress, 31556952, keccak256(bytes(name)), address(publicResolver), new bytes[](0), true, 0
+        );
+
         console.logBytes32(commitment);
         flrRegistrarController.commit(commitment);
-        
+
         console.log("Before Warp: block.timestamp = %s", block.timestamp);
         vm.warp(block.timestamp + 600);
         console.log("After Warp: block.timestamp = %s", block.timestamp);
@@ -50,61 +44,48 @@ contract Template is Script, DeployFNSAbstract {
 
         console.log("Price of name:%s => %s", name, totalPrice);
 
-        flrRegistrarController.register{ value: totalPrice }(
-            name,
-            broadcastAddress,
-            31556952,
-            keccak256(bytes(name)),
-            address(publicResolver),
-            new bytes[](0),
-            true,
-            0);
-        
-        (bytes32 nodeHash, uint tokenId) = nameWrapper.getFLRTokenId(name);
+        flrRegistrarController.register{value: totalPrice}(
+            name, broadcastAddress, 31556952, keccak256(bytes(name)), address(publicResolver), new bytes[](0), true, 0
+        );
+
+        (bytes32 nodeHash, uint256 tokenId) = nameWrapper.getFLRTokenId(name);
         console.logBytes32(nodeHash);
         console.log("tokenId=%s", tokenId);
 
         // Add subdomain
         console.log("[Script] --- setSubnodeRecord ---");
-        bytes32 subdomainNode = nameWrapper.setSubnodeRecord(
-            nodeHash,
-            "sub",
-            broadcastAddress,
-            address(publicResolver),
-            0,
-            0,
-            0);
+        bytes32 subdomainNode =
+            nameWrapper.setSubnodeRecord(nodeHash, "sub", broadcastAddress, address(publicResolver), 0, 0, 0);
         bytes32 subdomainNamehash = FNSNamehash.namehash("sub.test.flr");
         console.logBytes32(subdomainNamehash);
         console.logBytes32(subdomainNode);
         console.log("[SUBDOMAIN] ownerOf(\"sub.test\"): %s", nameWrapper.ownerOf(uint256(subdomainNode)));
-        
+
         (ISubdomainTracker.Data[] memory data, uint256 length) = subdomainTracker.getAll(uint256(nodeHash));
         console.log("SubdomainTracker length=%s", length);
-        for(uint i = 0; i < length; ++i){
+        for (uint256 i = 0; i < length; ++i) {
             console.log("data[%s]\n\tid=%s\n\towner=%s\n\t", i, data[i].id, data[i].owner);
             console.logString(data[i].label);
         }
 
         (IMintedDomainNames.Data[] memory data2, uint256 length2) = mintedDomainNames.getAll(broadcastAddress);
         console.log("MintedDomainNames length=%s", length2);
-        for(uint i = 0; i < length2; ++i){
+        for (uint256 i = 0; i < length2; ++i) {
             console.log("data2[%s]\n\tid=%s\n\texpiry=%s\n\t", i, data2[i].id, data2[i].expiry);
             console.logString(data2[i].label);
         }
 
         // Transfer to different owner
         bytes memory emptyBytes;
-        nameWrapper.safeTransferFrom(
-            broadcastAddress, DEPLOYER_ADDRESS, uint256(subdomainNode), 1, emptyBytes);
-        
+        nameWrapper.safeTransferFrom(broadcastAddress, DEPLOYER_ADDRESS, uint256(subdomainNode), 1, emptyBytes);
+
         (data, length) = subdomainTracker.getAll(uint256(nodeHash));
         console.log("SubdomainTracker length=%s", length);
-        for(uint i = 0; i < length; ++i){
+        for (uint256 i = 0; i < length; ++i) {
             console.log("data[%s]\n\tid=%s\n\towner=%s\n\t", i, data[i].id, data[i].owner);
             console.logString(data[i].label);
         }
-        
+
         // (IMintedDomainNames.Data[] memory data, uint256 length) = mintedDomainNames.getAll(broadcastAddress);
         // console.log("MintedDomainNames length=%s", length);
         // for(uint i = 0; i < length; ++i){
@@ -114,14 +95,14 @@ contract Template is Script, DeployFNSAbstract {
 
         (data2, length2) = mintedDomainNames.getAll(DEPLOYER_ADDRESS);
         console.log("MintedDomainNames length=%s", length2);
-        for(uint i = 0; i < length2; ++i){
+        for (uint256 i = 0; i < length2; ++i) {
             console.log("data2[%s]\n\tid=%s\n\texpiry=%s\n\t", i, data2[i].id, data2[i].expiry);
             console.logString(data2[i].label);
         }
 
         (data2, length2) = mintedDomainNames.getAll(broadcastAddress);
         console.log("MintedDomainNames length=%s", length2);
-        for(uint i = 0; i < length2; ++i){
+        for (uint256 i = 0; i < length2; ++i) {
             console.log("data2[%s]\n\tid=%s\n\texpiry=%s\n\t", i, data2[i].id, data2[i].expiry);
             console.logString(data2[i].label);
         }
@@ -130,7 +111,7 @@ contract Template is Script, DeployFNSAbstract {
         // string memory reverseName = publicResolver.name(reverseRegistrar.node(broadcastAddress));
         // console.log("NAME SET FOR: %s", broadcastAddress);
         // console.logString(reverseName);
-        
+
         // vm.stopBroadcast();
 
         // // Now set the name for the new owner
